chore: Separate nested unit tests from their abstract base class
Unit tests and coverage worked fine in Eclipse, however, no from
the command line. I'm guessing this has something to do with the
difference in versions for junit and/or Jacoco.

I separated the concrete, nested classes from the abstract base class
and that seemed to fix the problem.
